requirements.txt
goldmansachs.awm_genai
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
ipywidgets>=8.0.0
pdfplumber>=0.10.0
networkx>=3.1
matplotlib>=3.7.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0

